remove setup.sh just if apt-get was performed

+++ apps/guadalinex-nano-installer/trunk/setup.sh
@@ -8,5 +8,8 @@ wget http://gensys/$GNICONF -O /etc/gni-options
 # launch menu app
 /usr/sbin/gni
 
-# remove myself (/sbin/setup.sh)
-rm $0
+# remove myself (/sbin/setup.sh) if apt-get install was launched
+if [ $? -eq 0 ]
+then
+	rm $0
+fi
